user registration L15 C6.4 others all

+++ app/Http/Controllers/TasksController.php
@@ -151,9 +151,16 @@ class TasksController extends Controller
     {
         //
         // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
+        //$task = Task::findOrFail($id);
         // メッセージを削除
-        $task->delete();
+        //$task->delete();
+        
+        //Lesson 15Chapter 9.5 destroy
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
 
         // トップページへリダイレクトさせる
         return redirect('/');
